+++ src/test/java/act/route/RouterBenchmark.java
@@ -1,7 +1,26 @@
 package act.route;
 
+/*-
+ * #%L
+ * ACT Framework
+ * %%
+ * Copyright (C) 2014 - 2018 ActFramework
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
 import act.Act;
-import act.ActResponse;
 import act.BenchmarkBase;
 import act.app.ActionContext;
 import act.app.App;
@@ -29,7 +48,6 @@ import java.lang.reflect.Field;
 import java.net.URL;
 import java.util.Map;
 
-import static org.mockito.Mockito.mock;
 import static org.osgl.http.H.Method.GET;
 import static org.osgl.http.H.Method.POST;
 
@@ -69,7 +87,7 @@ public class RouterBenchmark extends BenchmarkBase {
 
     void osgl(H.Method method, String url, Object... args) {
         H.Request req = new MockRequest(config, method, url);
-        ctx = ActionContext.create(app, req, mock(ActResponse.class));
+        ctx = ActionContext.create(app, req, new MockResponse());
         if (args.length > 0) {
             url = S.fmt(url, args);
         }
@@ -167,7 +185,7 @@ public class RouterBenchmark extends BenchmarkBase {
 
     private void runTest(boolean osgl, boolean notFoundExpected, H.Method method, String url, Object... fmtArgs) {
         url = S.fmt(url, fmtArgs);
-        final int loop = 1000 * 10;
+        final int loop = 1000 * 100;
         if (osgl) {
             for (int i = 0; i < loop; ++i) {
                 try {